requirements.txt
python-vlc==3.0.12118
RPi.GPIO==0.7.1